Split admm module into submodules
