+++ scico/optimize/admm/__init__.py
@@ -1,0 +1,31 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2022 by SCICO Developers
+# All rights reserved. BSD 3-clause License.
+# This file is part of the SCICO package. Details of the copyright and
+# user license can be found in the 'LICENSE' file distributed with the
+# package.
+
+"""ADMM solver and auxiliary classes."""
+
+import sys
+
+# isort: off
+from ._subproblem import (
+    SubproblemSolver,
+    GenericSubproblemSolver,
+    LinearSubproblemSolver,
+    CircularConvolveSolver,
+)
+from ._admm import ADMM
+
+__all__ = [
+    "SubproblemSolver",
+    "GenericSubproblemSolver",
+    "LinearSubproblemSolver",
+    "CircularConvolveSolver",
+    "ADMM",
+]
+
+# Imported items in __all__ appear to originate in top-level linop module
+for name in __all__:
+    getattr(sys.modules[__name__], name).__module__ = __name__
